Update Message datestamp to double

+++ BlazorDungeonCrawler/BlazorDungeonCrawler.Database/Migrations/202310081652066_DBSetup.cs
@@ -103,7 +103,7 @@
                     {
                         Id = c.Guid(nullable: false),
                         Index = c.Int(nullable: false),
-                        Datestamp = c.Long(nullable: false),
+                        Datestamp = c.Double(nullable: false),
                         Text = c.String(),
                         Dungeon_Id = c.Guid(),
                     })
